training/requirements.txt
# Training dependencies
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
opencv-python>=4.8.0
tqdm>=4.65.0
numpy>=1.24.0
requests>=2.28.0
faiss-cpu>=1.7.4
onnx>=1.14.0
onnxruntime>=1.15.0
